Update Global.cs to use Properties
If these are not properties and are instead just values then the order in which things are declared in this class matter else you'll end up with empty values. I know this is just demo code and in real life these values should be validated, etc... but it would make things a little more error tolerant to have them as properties to begin with.